// projects/ng7-snack-bar/src/lib/snack-bar.enum.ts
export enum BarType {
    INFO = 'INFO',
    SUCCESS = 'SUCCESS',
    WARN = 'WARN',
    ERROR = 'ERROR',
}

export enum PositionType {
    TOP_LEFT = 'TOP_LEFT',
    TOP_RIGHT = 'TOP_RIGHT',
    BOTTOM_LEFT = 'BOTTOM_LEFT',
    BOTTOM_RIGHT = 'BOTTOM_RIGHT',
}

export enum TemplateType {
    CLASSIC = 'CLASSIC',
    DARK = 'DARK',
}

export enum GroupRuleType {
    NO_RULE = 'NO_RULE',
    BASED_ON_TYPE = 'BASED_ON_TYPE',
    BASED_ON_CONTENT = 'BASED_ON_CONTENT',
}

export enum SnackBarType {
  SHOW_SNACKBAR = 'SHOW_SNACKBAR',
  SET_SNACKBAR_IOPTIONS = 'SET_SNACKBAR_IOPTIONS',
  REMOVE_SNACKBAR_ON_CLOSE = 'REMOVE_SNACKBAR_ON_CLOSE',
  CLEAR_SNACKBAR = 'CLEAR_SNACKBAR',
  SET_CURRENT_URL = 'SET_CURRENT_URL',
}